Merged PR 18717: Add EnvironmentVariables to set AmsRouter Endpoint
For containerized environments we need the option to set the AMS Router
Endpoint via environment variables. This commit is just an example how
I implemented that option for first tests which worked fine with the
conatinerized AdsRouteConsoleApp. However I think there is an
appropriated .net library which should be used for parsing the passed
environment variables.